разбивка на модули, export & import
