+++ package.json
@@ -7,13 +7,14 @@
     "type": "git",
     "url": "git@github.com:htmlacademy-javascript/852387-kekstagram-24.git"
   },
+  "type": "module",
   "bugs": {
     "url": "https://github.com/htmlacademy-javascript/852387-kekstagram-24/issues"
   },
   "devDependencies": {
     "browser-sync": "2.26.14",
     "eslint": "7.18.0",
-    "eslint-config-htmlacademy": "4.0.0"
+    "eslint-config-htmlacademy": "4.1.0"
   },
   "license": "MIT",
   "dependencies": {},
